Revert "[기능 개선] 회원가입 컴포넌트 세분화"
This reverts commit d01bca21bb2b450ea7b675f66bae68b1182040de.
